+++ CHANGES.rst
@@ -35,6 +35,10 @@ Changelog
   value term of `IRedefineAttendeePosition.position_type` correctly when it was
   not in the base vocabulary.
   [gbastien]
+- Display `MeetingConfig` title in page title (displayed in web browser tab)
+  on faceted contexts (dashboard and in configuration) so user knows where he is
+  when using several tabs.
+  [gbastien]
 - Fixed `PMCategorizedObjectAdapter.can_view` when managing not viewable items
   because it could lead to unwanted users having access to some annexes by
   accessing it's URL directly.